+++ src/main/java/com/google/inject/internal/DynamicSingletonProvider.java
@@ -2,8 +2,8 @@ package com.google.inject.internal;
 
 import com.google.inject.Injector;
 
-import javax.inject.Inject;
-import javax.inject.Singleton;
+import com.google.inject.Inject;
+import com.google.inject.Singleton;
 import java.lang.annotation.Annotation;
 
 /**