+++ src/main/java/ru/vyarus/guice/ext/log/Log.java
@@ -1,6 +1,7 @@
 package ru.vyarus.guice.ext.log;
 
-import javax.inject.Scope;
+import com.google.inject.ScopeAnnotation;
+
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -11,7 +12,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 /**
  * Log annotation must be used on {@code org.slf4j.Logger} fields to automatically inject logger instance.
  */
-@Scope
+@ScopeAnnotation
 @Documented
 @Retention(RUNTIME)
 @Target(FIELD)