+++ src/test/java/ru/vyarus/guice/ext/generator/support/anchor/CtorAbstractClass.java
@@ -4,7 +4,7 @@ import com.google.inject.ProvidedBy;
 import com.google.inject.internal.DynamicSingletonProvider;
 import ru.vyarus.guice.ext.generator.support.aop.CustomAop;
 
-import javax.inject.Inject;
+import com.google.inject.Inject;
 
 /**
  * @author Vyacheslav Rusakov